Don't try to blindly reconnect on connections failures
On a replica set, when the master changes, the old master disconnects
all clients. This mean we have to trigger a cluster refresh on every
disconnect and don't try not to blindly reconnect.

In order to do that we propagate exceptions from `Node#ensure_connected`
and `Connection#with_connection`.

Before that, a non-safe session never caught up with a failover and all
subsequent inserts were silently lost.

The issue was present in 1.4.2 and, as of today, is still present from 1.5.0
throughout 2.0.6.

See https://github.com/mongoid/moped/pull/153.